add teca_string_util
collects string processing codes into a utility module.
